# openapi.yaml
openapi: 3.0.0
paths:
...
  /user/community:
    post:
      operationId: CommunityController_createCommunity
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommunityDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/CommunityResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Create new community
      tags: &ref_1
        - Community
    get:
      operationId: CommunityController_getAllCommunities
      parameters:
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
        - name: type
          required: true
          in: query
          schema:
            type: string
        - name: search
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful paginated response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CommunityResponseDto'
                    description: Array of data items
                  paging:
                    type: object
                    description: Pagination metadata
                    properties:
                      total:
                        type: number
                        example: 100
                        description: Total number of items
                      page:
                        type: number
                        example: 1
                        description: Current page number
                      limit:
                        type: number
                        example: 10
                        description: Items per page
                      totalPages:
                        type: number
                        example: 10
                        description: Total number of pages
                    required:
                      - total
                      - page
                      - limit
                      - totalPages
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
                  - data
                  - paging
      summary: Get all communities with pagination
      tags: *ref_1
  /user/community/{communityId}:
    get:
      operationId: CommunityController_getCommunity
      parameters:
        - name: communityId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/CommunityResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Get community by ID
      tags: *ref_1
    patch:
      operationId: CommunityController_updateCommunity
      parameters:
        - name: communityId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommunityDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/CommunityResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Update community profile
      tags: *ref_1
    delete:
      operationId: CommunityController_deleteCommunity
      parameters:
        - name: communityId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: object
                    nullable: true
                    example: null
                    description: No data returned
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Delete community
      tags: *ref_1
  /user/community/{communityId}/follow:
    post:
      operationId: CommunityController_followCommunity
      parameters:
        - name: communityId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: object
                    nullable: true
                    example: null
                    description: No data returned
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Follow community
      tags: *ref_1
    delete:
      operationId: CommunityController_unfollowCommunity
      parameters:
        - name: communityId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: object
                    nullable: true
                    example: null
                    description: No data returned
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Unfollow community
      tags: *ref_1
...
  /user/post/community/{communityId}:
    get:
      operationId: PostController_getPostsByCommunity
      parameters:
        - name: communityId
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
      responses:
        '200':
          description: Successful paginated response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostResponseDto'
                    description: Array of data items
                  paging:
                    type: object
                    description: Pagination metadata
                    properties:
                      total:
                        type: number
                        example: 100
                        description: Total number of items
                      page:
                        type: number
                        example: 1
                        description: Current page number
                      limit:
                        type: number
                        example: 10
                        description: Items per page
                      totalPages:
                        type: number
                        example: 10
                        description: Total number of pages
                    required:
                      - total
                      - page
                      - limit
                      - totalPages
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
                  - data
                  - paging
      summary: Get posts by community
      tags: *ref_2
...
info:
  title: API Docs
  description: Custom Swagger UI in NestJS
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CommunityResponseDto:
      type: object
      properties:
        id:
          type: string
          example: comm_abc123
        name:
          type: string
          example: K-Pop Lovers
        type:
          type: string
          example: GROUP
          enum:
            - SOLO
            - GROUP
        ownerId:
          type: string
          example: user_xyz789
        description:
          type: object
          example: A community for K-Pop enthusiasts
          nullable: true
        avatarUrl:
          type: object
          example: https://example.com/avatar.jpg
          nullable: true
        backgroundUrl:
          type: object
          example: https://example.com/background.jpg
          nullable: true
        followerCount:
          type: number
          example: 1250
        postCount:
          type: number
          example: 430
        isActive:
          type: boolean
          example: true
        createdAt:
          format: date-time
          type: string
          example: '2024-01-15T08:30:00Z'
        updatedAt:
          format: date-time
          type: string
          example: '2024-01-20T14:45:00Z'
      required:
        - id
        - name
        - type
        - ownerId
        - description
        - avatarUrl
        - backgroundUrl
        - followerCount
        - postCount
        - isActive
        - createdAt
        - updatedAt
    CreateCommunityDto:
      type: object
      properties: {}
    UpdateCommunityDto:
      type: object
      properties: {}
...
    PostMediaResponseDto:
      type: object
      properties:
        url:
          type: string
          example: https://example.com/media/image1.jpg
        order:
          type: number
          example: 0
      required:
        - url
        - order
    PostResponseDto:
      type: object
      properties:
        id:
          type: string
          example: post_abc123
        authorId:
          type: string
          example: user_xyz789
        communityId:
          type: object
          example: comm_def456
          nullable: true
        content:
          type: string
          example: This is an amazing post about K-Pop!
        media:
          type: array
          items:
            $ref: '#/components/schemas/PostMediaResponseDto'
        likeCount:
          type: number
          example: 245
        viewCount:
          type: number
          example: 1520
        commentCount:
          type: number
          example: 34
        isDeleted:
          type: boolean
          example: false
        createdAt:
          format: date-time
          type: string
          example: '2024-01-15T08:30:00Z'
        updatedAt:
          format: date-time
          type: string
          example: '2024-01-20T14:45:00Z'
      required:
        - id
        - authorId
        - communityId
        - content
        - media
        - likeCount
        - viewCount
        - commentCount
        - isDeleted
        - createdAt
        - updatedAt
...